delete walletconnect model, delete addTransactionsToApprove, delete walletConnectGetAllRequests, delete walletConnectGetRequest, delete removeAllDeliveredNotifications
